Enable `RTLD_DEEPBIND` by default, add workaround for musl and FreeBSD
We allow the vast majority of functionality for musl and FreeBSD, the
only two systems that do not have `RTLD_DEEPBIND` functionality.  Our
method enables usage of most combinations of BLAS libraries, with the
exception of a 32-bit BLAS and a 64-bit BLAS that uses no suffix on its
symbols.