Add `f2c` autodetection, test against Accelerate on macOS
Accelerate has a historically troublesome calling convention that stems
from its use of `f2c`, resulting in return values that should be `float`
instead returning as `double`.  This results in mangled values in code
that expects the return value to be `float` and must be corrected in
wrapper stub functions.  The same holds true for `float complex` and
`double complex` types.  This PR adds these stub functions (referred to
as `f2c` functions) for the list of functions defined in
`FLOAT32_FUNCS`, `COMPLEX64_FUNCS` and `COMPLEX128_FUNCS`, currently
curated in `ext/gensymbol/generate_func_list.sh`.

We autodetect the presence of an `f2c`-style BLAS library by calling
`sdot_()` with some known inputs and looking at the output.  In the
event that the return value is properly interpretable if we expect a
`double` return value, we engage our `floatret` wrapper functions.  This
capability is only available on `x86_64` as it's not known to be
necessary anywhere else.