Add "Footgun API"
This adds the `lbt_set_forward()` and `lbt_get_forward()` APIs, which
allow directly setting the BLAS API functions that a user might be
interested in wrapping.

It also adds `lbt_set_default_function` and provides a fairly useless
(but at least not fatal) default value for the default function so that
if users don't want to seqfault when calling an uninitialized function,
they won't.