Add `active_forwards` to `lbt_libinfo_t`
This adds a new field to the `lbt_libinfo_t` structure that allows
determining which BLAS/LAPACK library calls are being satisfied by which
loaded libraries.  By inspecting the `active_forwards` field it will be
possible to, for example, walk over all loaded libraries and determine
which one the `dgemm_64_` call is provided by, which may be useful for
applications that, for instance, tune threading thresholds based on
which BLAS provider is being used.