Mark stack as non executable
The linker is unable to detect that executable stack is not required and so marks
libblastrampoline.so as requiring it. Pass `-Wl,-z,noexecstack` to
ensure that the stack is marked as not executable. This improves security and
allows running on systems where SELinux has been configured to disallow
executable stacks.