Enable complex retstyle detection on windows as well
I previously thought that the Intel standard for complex return values
on windows was the default for the platform, and so disabled the shims
to work around it. However, it turns out that gfortran uses a consistent
ABI across platforms, and we want to standardize on that, so let's
enable this again. This should fix the memory corruption seen when
calling MKL's `cdotc_64` on `x86_64` Windows.