Add support for ILP64 Accelerate
The new Accelerate released in macOS v13.3 provides two new interfaces;
an upgraded LAPACK for the LP64 interface, and an all-new ILP64
interface (that uses the same upgraded LAPACK).  These symbols are
available from Accelerate with the suffix `$NEWLAPACK` and
`$NEWLAPACK$ILP64`, respectively.

Unfortunately, this is not a "true" suffix, as Apple has decided to drop
the trailing underscore from the typical F77 names, meaning that a
symbol such as `dgemm_` gets mangled to `dgemm$NEWLAPACK`, whereas a
CBLAS symbol such as `cblas_zdotc_sub` gets mangled to
`cblas_zdotc_sub$NEWLAPACK`.  This means that we need to selectively
erase the trailing underscore from some symbols when applying this
Accelerate suffix.

To do this, we add a new feature, enabled by default only on Apple
builds, called `SYMBOL_TRIMMING`, which allows a `suffix_hint` to
contain the ASCII "substitution character" `0x1a` as the first character
of the suffix hint to mean "remove a trailing underscore when applying
this suffix".

To make dealing with suffix hints easier for command-line users, these
suffix hints are available for use in `LBT_BACKING_LIBS` by listing
libraries separated by suffix hints with an exclamation point,  e.g.
`libname!suffix`.