Prevent setting infinite loopback forwards
Some libraries, such as `libLAPACK` re-export their imported symbols,
presumably in an effort to make it easy for users to link against just
`libLAPACK` and get the BLAS symbols for free.  However, when LAPACK_jll
has linked against LBT as its backing BLAS library, we end up with an
infinite loop, as LBT then sets its BLAS symbols to its own trampoline.

This commit adds a special-case to our forwarding logic which explicitly
checks for such a loop, and silently refuses to perform such a forward.
This has the happy effect of filtering out direct loops, and makes it
much easier to mix and match libraries that themselves depend on LBT.