Support `libdir` and `includedir` variables
These are the standard GNU make variables that distributions expect
to be able to set
(https://www.gnu.org/prep/standards/html_node/Directory-Variables.html).
Since Julia supports the same variables and passes them during the build,
this ensures that libblastrampoline.so is installed to the right directory
(in particular to usr/lib64 where appropriate).