Allow for overriding workarounds even when autodetection is unavailable
If a user wants to use MKL on i686, we unfortunately need some
workarounds for certain function calls.  Until we come up with an
autodetection strategy that works on i686 as well as x86_64, we can make
use of the new environment variable-based overrides to make things work.
Let's test that on i686 on CI.