Tell `armv7l` assembler this is a constant
Because armv7 opcodes are only 32 bits, `mov` can only move immediates
that are 8 bits or less.  There's some wiggle room with shifting and
whatnot, but in full generality you need to actually perform a load from
a memory address.  The assembler will insert the immediate into a
literal pool, but that pool has to be <4KB away from the current `pc` at
the time of load, so we add `.ltorg` to flush the literal pool at the
end of every function.