Add fallback search for extra-underscore-suffixed symbols
MKL v2024 has ILP64-suffixed symbols, but the suffix applied to FORTRAN
symbols (`64`, mapping `dgemm_` to `dgemm_64`) is different from the
suffix applied to non-FORTRAN symbols (`_64`, mapping `cblas_dgemm` to
`cblas_dgemm_64`).  This wreaks havoc with LBT, which assumes that all
symbols undergo the same name transformation.  To work around this, we
add a fallback path to our symbol forwarding routine; if a symbol does
not exist in a library, we look for the same symbol but with an
underscore in front of the symbol name. Because this all happens only
after we have already found `isamax_` and `dpotrf_` in
`autodetect_symbol_suffix()`, we have some measure of assurance that we
will not be blindly guessing ridiculous symbol names.