Fix symbol autodetection to ignore our own symbols
On Linux, dlsym() will search the entire dependency tree of the library
we load, so if we are trying to load a library that is linked against
LBT, then dlsym() could give us back the symbol in LBT instead of the
one actually in the target library.

To prevent this, we need to check to see if the returned symbol is ours,
and ignore it if it is ours.